financial-document-analyzer-debug/requirements.txt
fastapi
uvicorn
pydantic
crewai
crewai-tools
python-dotenv
